backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
redis==5.0.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
prometheus-client==0.19.0
python-dotenv==1.0.0
google-generativeai==0.8.6
